[trunk] The last warning of compiler was eliminated. It was connected with size_t type and unsigned int. In future we will associate size_t with unsigned int (because all functions which work with sizes of objects are written this way).

+++ trunk/sources/funcsim/memory.h
@@ -269,7 +269,7 @@ public:
     /* The constant member function. Return size of Byteline */
     unsigned int getSizeOfLine() const
     {
-        return ( *byte_line).size();
+        return (unsigned int)( *byte_line).size();
     }
 
     OutputFormat getOutputFormat() const
